Rename higher order components.
